build(monorepo-scripts): automate contract addresses and artifacts
whenever the protocol package changes, generate and commit to origin the updated source code for
contract-addresses and contract-artifacts

+++ packages/monorepo-scripts/artifacts/build.js
@@ -7,39 +7,18 @@ const CONTRACTS_DIR = path.join(__dirname, '..', '..', 'protocol', 'build', 'con
 const ARTIFACTS_DIR = path.join(__dirname, '..', '..', 'contract-artifacts', 'artifacts');
 const ARTIFACTS_DIR_AUX = path.join(__dirname, '..', '..', 'contract-artifacts', 'artifacts-aux');
 
-const excludedContracts = [
-    'ERC20.json',
-    'IERC20.json',
-    'Migrations.json',
-    'Ownable.json',
-    'SafeMath.json',
-];
+const excludedContracts = ['ERC20', 'IERC20', 'Migrations', 'Ownable', 'SafeMath'];
 
-if (!fs.existsSync(CONTRACTS_DIR)) {
-    console.error('Please use truffle to compile your contracts first');
-    process.exit(1);
-}
-
-if (!fs.existsSync(ARTIFACTS_DIR)) {
-    fs.mkdir(ARTIFACTS_DIR);
-}
+const abiSwapContracts = ['AdjustSupply', 'BilateralSwap', 'DividendComputation', 'JoinSplit'];
 
 const cherrypickContractJson = async (filename) => {
-    const content = JSON.parse(await fs.readFile(path.join(CONTRACTS_DIR, filename), 'utf-8'));
-    if (filename === 'AdjustSupply') {
-        const contractInterface = JSON.parse(await fs.readFile(path.join(CONTRACTS_DIR, 'AdjustSupplyInterface.sol')));
-        content.abi = contractInterface.abi;
-    }
-    if (filename === 'BilateralSwap') {
-        const contractInterface = JSON.parse(await fs.readFile(path.join(CONTRACTS_DIR, 'BilateralSwapInterface.sol')));
-        content.abi = contractInterface.abi;
-    }
-    if (filename === 'DividendComputation') {
-        const contractInterface = JSON.parse(await fs.readFile(path.join(CONTRACTS_DIR, 'DividendComputationInterface.sol')));
-        content.abi = contractInterface.abi;
-    }
-    if (filename === 'JoinSplit') {
-        const contractInterface = JSON.parse(await fs.readFile(path.join(CONTRACTS_DIR, 'JoinSplitInterface.sol')));
+    const filepath = path.join(CONTRACTS_DIR, filename);
+    const content = JSON.parse(await fs.readFile(filepath, 'utf-8'));
+
+    const filenameWithoutExtension = filename.replace('.json', '');
+    if (abiSwapContracts.includes(filenameWithoutExtension)) {
+        const contractInterfacePath = path.join(CONTRACTS_DIR, `${filenameWithoutExtension}Interface.json`);
+        const contractInterface = JSON.parse(await fs.readFile(contractInterfacePath));
         content.abi = contractInterface.abi;
     }
 
@@ -58,26 +37,37 @@ const isDuplicate = async (filename) => {
         return false;
     }
     const previousContent = JSON.parse(await fs.readFile(path.join(ARTIFACTS_DIR, filename), 'utf-8'));
-    return content.abi === previousContent.abi
-        && content.bytecode === previousContent.bytecode
-        && content.compiler === previousContent.compiler
-        && content.deployedBytecode === previousContent.deployedBytecode
-        && content.schemaVersion === previousContent.schemaVersion
-        && content.updatedAt === previousContent.updatedAt;
+    return (
+        content.abi === previousContent.abi &&
+        content.bytecode === previousContent.bytecode &&
+        content.compiler === previousContent.compiler &&
+        content.deployedBytecode === previousContent.deployedBytecode &&
+        content.schemaVersion === previousContent.schemaVersion &&
+        content.updatedAt === previousContent.updatedAt
+    );
 };
 
-const compileArtifacts = async () => {
-    await fs.remove(ARTIFACTS_DIR_AUX);
-    await fs.mkdir(ARTIFACTS_DIR_AUX);
+const buildArtifacts = async () => {
     try {
+        if (!fs.existsSync(CONTRACTS_DIR)) {
+            throw new Error('Please use truffle to compile your contracts first');
+        }
+        if (!fs.existsSync(ARTIFACTS_DIR)) {
+            fs.mkdir(ARTIFACTS_DIR);
+        }
+        await fs.remove(ARTIFACTS_DIR_AUX);
+        await fs.mkdir(ARTIFACTS_DIR_AUX);
+
         let filenames = await fs.readdir(CONTRACTS_DIR);
-        filenames = filenames.filter(filename => {
-            return !filename.endsWith('Test.json') && !excludedContracts.includes(filename);
+        filenames = filenames.filter((filename) => {
+            const filenameWithoutExtension = filename.replace('.json', '');
+            return !filenameWithoutExtension.endsWith('Test') && !excludedContracts.includes(filenameWithoutExtension);
         });
+
         for (let i = 0; i < filenames.length; i += 1) {
             const filename = filenames[i];
             /* eslint-disable no-await-in-loop */
-            if (!await isDuplicate(filename)) {
+            if (!(await isDuplicate(filename))) {
                 const artifact = await cherrypickContractJson(filename);
                 await fs.writeFile(path.join(ARTIFACTS_DIR_AUX, filename), JSON.stringify(artifact, null, 4));
             }
@@ -88,7 +78,8 @@ const compileArtifacts = async () => {
         console.log('Successfully generated artifacts in the contract-artifacts package');
     } catch (err) {
         console.error('Failed with error:', err);
+        process.exit(1);
     }
 };
 
-compileArtifacts();
+buildArtifacts();
